add Degree NValues
